+++ src/main/java/Etf/comment/controller/CommentRestController.java
@@ -1,12 +1,13 @@
 package Etf.comment.controller;
 
 
-import Etf.comment.dto.CommentResponse;
+import Etf.comment.dto.CommentCreateRequest;
+import Etf.comment.dto.CommentUpdateRequest;
 import Etf.comment.dto.CommentsPageList;
+import Etf.comment.serviece.CommentLikeService;
 import Etf.comment.serviece.CommentService;
 import Etf.loginUtils.LoginMember;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -14,18 +15,52 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v1/comments")
 public class CommentRestController {
 
     private final CommentService commentService;
+    private final CommentLikeService commentLikeService;
 
-    @GetMapping("/users/comments")
+    @GetMapping
     public ResponseEntity<CommentsPageList> readAllComment(@LoginMember String loginId, @PageableDefault(page = 0, size = 20, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable, @RequestParam(name = "etf_id") Long etfId){
         CommentsPageList commentsPageList = commentService.readAll(loginId, pageable, etfId);
         return ResponseEntity.status(HttpStatus.OK).body(commentsPageList);
+    }
+
+    //좋아요 토글
+    @PostMapping("/{commentId}/likes/{userId}")
+    public void toggleLike(
+            @PathVariable Long commentId,
+            @PathVariable Long userId) {
+        commentLikeService.toggleLike(commentId, userId);
+    }
+
+    //Comment Create
+    @PostMapping
+    public void createComment(
+            @RequestBody CommentCreateRequest commentCreateRequest) {
+        commentService.create(commentCreateRequest);
+
+    }
+
+    //Comment Update
+    @PutMapping("/{commentId}/{userId}")
+    public void updateComment(
+            @PathVariable Long commentId,
+            @PathVariable Long userId,
+            @RequestBody CommentUpdateRequest commentUpdateRequest) {
+        commentService.update(commentId, userId, commentUpdateRequest);
+
+    }
+
+    //Comment Soft Delete
+    @DeleteMapping("/{commentId}/{userId}")
+    public void deleteComment(
+            @PathVariable Long commentId,
+            @PathVariable Long userId) {
+        commentService.delete(commentId, userId);
+
     }
 }
