+++ src/main/java/Etf/comment/dto/CommentRequest.java
@@ -1,6 +1,5 @@
 package Etf.comment.dto;
 
-import Etf.comment.domain.Comment;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
@@ -10,6 +9,6 @@ public record CommentRequest(
         @NotNull@Size(max = 1000, min = 2)
         String content,
         @NotNull
-        Long parentId
+        String UserName
 ) {
 }
