--- src/main/java/Etf/comment/dto/CommentRequest2.java
@@ -1,14 +1,0 @@
-package Etf.comment.dto;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-
-public record CommentRequest2(
-        @NotNull
-        Long etfId,
-        @NotNull@Size(max = 1000, min = 2)
-        String content,
-        @NotNull
-        String UserName
-) {
-}
