+++ src/main/java/Etf/comment/serviece/CommentLikeService.java
@@ -5,8 +5,8 @@ import Etf.comment.domain.Comment;
 import Etf.comment.domain.CommentLike;
 import Etf.comment.repository.CommentLikeRepository;
 import Etf.comment.repository.CommentRepository;
-import EtfRecommendService.user.User;
-import EtfRecommendService.user.UserRepository;
+import Etf.user.User;
+import Etf.user.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
